Almost all of the basic features have been added
murkyvault.py:

Almost all of the basic features have been added.

The user menu has been updated to add new filtering options by User Name, Application or Website and Email.

Functions have been created to allow passwords to be encrypted when inserted into the database and, in turn, to be decrypted to display them as plain text when the user requests them.