Customizing the output tables and adding secondary user options.
murkyvault.py:

The visual appearance of the data output of the tables by the terminal has been customized.

Secondary options have been added at the end of each main option, so that the user can choose whether to repeat the same option with a new parameter or choose another of those available in the menu.