New feature: Delete Passwords
murkyvault.py:

New feature added:

A new function has been created which allows the user to delete a row from the table indicating the name of the Site or APP as a reference for the MySQL query.
The user will be asked for confirmation showing him the values ​​that that row contains before being eliminated.

The visual appearance of the main options menu has also been modified.