New feature: Modify Row
murkyvault.py:

A new function has been added that allows the user to modify all the fields of a row indicating the Website or APP as a parameter for the MySQL query.

The user will be asked for confirmation showing him the current fields of the row he is about to modify.

The user menu has been modified by changing the input parameters from Letters to Numbers.