New feature: Modify Master Password
murkyvault.py:

Added a new feature to modify the master password in case the user wants to change it.

For this process, the user will be asked to verify that he knows the previous password and if this is true, it will be updated with a new value entered by keyboard.