New feature: Query related to Class
murkyvault.py:

A new column has been added to the user's password table to allow a group to be associated with each of the records in the table itself.

In addition, the feature has been implemented to be able to consult the user's passwords indicating the name of the associated Group as the search parameter.

Fixed small string numbering bugs.