Database check or creation
murkyvault.py:

A new feature has been added that allows the creation of the database necessary for the operation of the application if it is not previously created in the user's MySQL Server.

Also, since the user can create several different databases, the current database is now displayed at the top of the user menu.