Query columns available to be able to filter passwords and visual changes have been added
murkyvault.py:

Several features have been added to be able to query the available columns in the database to filter further passwords. Example: Websites or APPs, Emails, Usernames and Groups.

In turn, the visual aspect of the main menu has been modified and the menu action numbers have been reorganized to apply the new features.