Start of web application development
murkyvault.py:

Moved the murkyvault.py program to a new folder to categorize it under the Terminal tag.

index.php, options.php, redirect.php and showAllPasswords.php (and respective CSS properties):

The process of creating a web application that provides the same features as the Terminal application has been initialized.

For this, the previous Project-WIRENET code design has been based to provide a minimalist interface with similar characteristics.

Several PHP files and their respective CSS have been created, which will be modified in the future.

In this snapshot, the option of initializing a connection with the Database and displaying all the passwords it contains is allowed without, yet, any filter.