New feature for the Web APP: AddPassword.
addPassword.php and related styles:

Added the ability to add passwords through a web application form. A confirmation has also been added for the user to decide whether or not to add more passwords, as well as the necessary styles to give the page the visual appearance of the rest of the project.