New feature added: Field query with filters applied.
Various PHP and CSS files:

Various PHP pages have been created with their respective CSS styles to add the feature of making queries by applying various Website or App, email, user or group filters.

It is allowed to relaunch the same option at the end of the query if the user wishes, otherwise the options menu will be displayed again.