New feature added: Column listing and record deletion.
Various PHP and CSS files:

Several PHP pages have been created with their respective CSS styles to create a list of columns where the user can consult which Sites or APPs are stored in the DB, as well as emails, users, etc. To later filter by these.

Likewise, another feature has been created to be able to delete a complete record indicating the name of the Website or APP as the query order.