New feature added: modifyRow
modifyRow.php and respective CSS file:

A new page has been created that allows the user to indicate a record to be modified, subsequently indicating the new values ​​that will be updated in the respective fields of the old record.

In turn, the associated styles have been created, the current fields of that record will be shown to the user prior to their modification to prevent possible accidental deletions.